Move isIndirect to static method of WeaponType
